pyproject.toml
[project]
name = "job-search-agent"
version = "0.1.0"
description = "Job search agent using LangChain Deep Agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core
    "deepagents>=0.1.0",
    "langchain>=0.3.0",
    "langchain-deepseek>=0.1.0",
    "langgraph>=0.2.0",

    # Search APIs
    "tavily-python>=0.5.0",
    "httpx>=0.27.0",

    # PDF parsing
    "pypdf>=4.0.0",

    # Utilities
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "markdownify>=0.13.0",

    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.17",

    # Database
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.2.0",
    "alembic>=1.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
